@@ -72,26 +71,25 @@ repositories {
 }
 
 gradlePlugin {
+    website.set("https://github.com/IntershopCommunicationsAG/${project.name}")
+    vcsUrl.set("https://github.com/IntershopCommunicationsAG/${project.name}")
     plugins {
-        register("ArchitectureReportPlugin") {
+        create("ArchitectureReportPlugin") {
             id = "com.intershop.gradle.architectural.report"
             implementationClass = "com.intershop.tool.architecture.report.plugin.ArchitectureReportPlugin"
             displayName = project.displayName
             description = project.description
+            tags.set(listOf("intershop", "validation", "analysis"))
         }
     }
 }
 
-pluginBundle {
-    website = "https://github.com/IntershopCommunicationsAG/${project.name}"
-    vcsUrl = "https://github.com/IntershopCommunicationsAG/${project.name}"
-    description = project.description
-    tags = listOf("intershop", "validation", "analysis")
-}
-
 java {
-    sourceCompatibility = JavaVersion.VERSION_11
-    targetCompatibility = JavaVersion.VERSION_11
+    sourceCompatibility = JavaVersion.VERSION_17
+    targetCompatibility = JavaVersion.VERSION_17
+
+    withJavadocJar()
+    withSourcesJar()
 }
 
 // set correct project status
@@ -110,16 +108,16 @@ tasks {
 
     register("generateResources") {
         // Generate properties file with plugin version to access this information in plugin itself
-        val versionPropertyFile = file("${buildDir}/generated/version.properties")
+        val versionPropertyFile = project.layout.buildDirectory.file("generated/version.properties")
         outputs.file(versionPropertyFile)
         doLast {
-            mkdir(versionPropertyFile.parentFile)
-            versionPropertyFile.writeText("version=${project.version}")
+            mkdir(versionPropertyFile.get().asFile.parentFile)
+            versionPropertyFile.get().asFile.writeText("version=${project.version}")
         }
     }
 
     withType<KotlinCompile>  {
-        kotlinOptions.jvmTarget = JavaVersion.VERSION_11.toString()
+        kotlinOptions.jvmTarget = JavaVersion.VERSION_17.toString()
     }
 
     withType<ProcessResources> {
